services/ingestion/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
uvloop==0.19.0
httptools==0.6.1
pydantic==2.7.0
pydantic-settings==2.3.0
httpx==0.27.0
structlog==24.2.0
python-multipart==0.0.9
pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
pytesseract==0.3.10
anthropic==0.25.0
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
sentry-sdk[fastapi]==2.5.0
pytest==8.2.0
pytest-asyncio==0.23.6
ruff==0.4.4
